backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
flask_bcrypt
flask_jwt_extended
sqlite3-binary
logging